requirements.txt
fastapi
pydantic
twilio
deta
attrs
python-decouple